Derek development

+++ app/src/main/java/com/example/pet_universe/database/ListingDatabaseDao.kt
@@ -34,5 +34,5 @@ interface ListingDatabaseDao {
     fun getListingById(listingId: Long): Listing?
 
     @Query("UPDATE listing_table SET title = :title, price = :price, description = :description, meeting_location = :meetingLocation, category = :category, imageUrls = :imageUrls WHERE id = :id")
-    fun update(id: Long, title: String, price: Double, description: String, meetingLocation: String, category: String, imageUrls: List<String>)
+    fun update(id: Long, title: String, price: Int, description: String, meetingLocation: String, category: String, imageUrls: List<String>)
 }
